Adds an experiment selector to the new run page. Needs tests
